pinterest_login/requirements.txt
browser-use==0.7.5
crewai==0.177.0
langchain-openai==0.3.32
playwright==1.55.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.0.0